+++ GR_AppInit.lua
@@ -29,6 +29,7 @@ GuildRecruiter = {}
 _G['GuildRecruiter'] = GuildRecruiter
 
 -- Constant Variables
+DEFAULT_GUILD_WELCOME = 'Welcome NAME to GUILDNAME!'
 GR_VERSION_INFO = 'Guild Recruiter v'..GRADDON.version
 SCAN_WAIT_TIME = 3
 PLAYER_PROFILE = UnitName('player')..' - '..GetRealmName()